requirements.txt
streamlit>=1.28.0
openai>=1.3.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=0.8.0
pandas>=2.1.0
plotly>=5.17.0
streamlit-authenticator>=0.2.0
streamlit-chat>=0.1.0